Refactor perform functions
