+++ packages/destination-actions/src/destinations/blackbaud-raisers-edge-nxt/createGift/index.ts
@@ -1,4 +1,4 @@
-import { ActionDefinition, IntegrationError } from '@segment/actions-core'
+import { ActionDefinition, InputField, IntegrationError } from '@segment/actions-core'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
 import {
@@ -6,10 +6,9 @@ import {
   perform as performCreateOrUpdateIndividualConstituent
 } from '../createOrUpdateIndividualConstituent'
 import { BlackbaudSkyApi } from '../api'
-import { Gift, GiftAcknowledgement, GiftReceipt } from '../types'
-import { dateStringToFuzzyDate } from '../utils'
+import { buildConstituentPayloadFromGiftPayload, buildGiftDataFromPayload } from '../utils'
 
-const fields = {
+const fields: Record<string, InputField> = {
   acknowledgement: {
     label: 'Acknowledgement',
     description: 'The gift acknowledgement.',
@@ -170,141 +169,31 @@ const fields = {
   }
 }
 
-// add constituent fields
 Object.keys(createOrUpdateIndividualConstituentFields).forEach((key: string) => {
   fields['constituent_' + key] = {
     ...createOrUpdateIndividualConstituentFields[key],
-    // prepend "Constituent" to label for clarity
-    label: 'Constituent ' + createOrUpdateIndividualConstituentFields[key].label,
-    constituentField: true
+    label: 'Constituent ' + createOrUpdateIndividualConstituentFields[key].label
   }
 })
 
 const perform = async (request, { payload }) => {
-  // check if request includes fields to create or update a constituent
-  // if so, append them to a new payload
-  const constituentPayload = {}
-  Object.keys(payload).forEach((key: string) => {
-    if (fields[key].constituentField) {
-      // only append non-empty fields/objects
-      if (payload[key] && (typeof payload[key] !== 'object' || Object.keys(payload[key]).length > 0)) {
-        const excludedConstituentFields = ['constituent_lookup_id']
-        if (!excludedConstituentFields.includes(key)) {
-          constituentPayload[key.substring('constituent_'.length)] = payload[key]
-        }
-      }
-    }
-  })
+  const constituentPayload = buildConstituentPayloadFromGiftPayload(payload)
 
   let constituentId = payload.constituent_id
   if (Object.keys(constituentPayload).length > 0) {
-    // create or update constituent
     const createOrUpdateIndividualConstituentResponse = await performCreateOrUpdateIndividualConstituent(request, {
       payload: constituentPayload
     })
     constituentId = createOrUpdateIndividualConstituentResponse.id
   } else if (constituentId === undefined) {
-    // constituent_id is required to create a gift
-    // no constituent_id, throw an error
-    throw new IntegrationError(`Missing constituent_id value`, 'MISSING_REQUIRED_FIELD', 400)
+    throw new IntegrationError('Missing constituent_id value', 'MISSING_REQUIRED_FIELD', 400)
   }
 
   const blackbaudSkyApiClient: BlackbaudSkyApi = new BlackbaudSkyApi(request)
 
-  // data for gift call
-  const giftData: Gift = {
-    constituent_id: constituentId,
-    // hardcode is_manual
-    is_manual: true
-  }
-  const simpleGiftFields = [
-    'amount',
-    'date',
-    'gift_status',
-    'is_anonymous',
-    'lookup_id',
-    'post_date',
-    'post_status',
-    'subtype',
-    'type'
-  ]
-  simpleGiftFields.forEach((key: string) => {
-    if (payload[key] !== undefined) {
-      giftData[key] = payload[key]
-    }
-  })
+  const giftData = buildGiftDataFromPayload(constituentId, payload)
 
-  // default date
-  giftData.date = giftData.date || new Date().toISOString()
-
-  // create acknowledgements array
-  if (payload.acknowledgement) {
-    const acknowledgementData: GiftAcknowledgement = {
-      status: payload.acknowledgement.status || 'NEEDSACKNOWLEDGEMENT'
-    }
-    if (
-      acknowledgementData.status !== 'NEEDSACKNOWLEDGEMENT' &&
-      acknowledgementData.status !== 'DONOTACKNOWLEDGE' &&
-      payload.acknowledgement.date
-    ) {
-      acknowledgementData.date = payload.acknowledgement.date
-    }
-    giftData.acknowledgements = [acknowledgementData]
-  }
-
-  // create gift splits array
-  giftData.gift_splits = [
-    {
-      amount: giftData.amount,
-      fund_id: payload.fund_id
-    }
-  ]
-
-  // create payments array
-  giftData.payments = [
-    {
-      payment_method: payload.payment_method
-    }
-  ]
-
-  // fields for check gifts
-  if (giftData.payments[0].payment_method === 'PersonalCheck') {
-    giftData.payments[0].check_number = payload.check_number
-    if (payload.check_date) {
-      const checkDateFuzzyDate = dateStringToFuzzyDate(payload.check_date)
-      if (checkDateFuzzyDate) {
-        giftData.payments[0].check_date = checkDateFuzzyDate
-      }
-    }
-  }
-
-  // default post date
-  if ((giftData.post_status === 'NotPosted' || giftData.post_status === 'Posted') && !giftData.post_date) {
-    giftData.post_date = payload.date
-  }
-
-  // create receipts array
-  if (payload.receipt) {
-    const receiptData: GiftReceipt = {
-      status: payload.receipt.status || 'NEEDSRECEIPT'
-    }
-    if (receiptData.status === 'RECEIPTED' && payload.receipt.date) {
-      receiptData.date = payload.receipt.date
-    }
-    giftData.receipts = [receiptData]
-  }
-
-  // fields for recurring gifts
-  if (giftData.type === 'RecurringGift') {
-    giftData.recurring_gift_schedule = payload.recurring_gift_schedule
-  } else if (giftData.type === 'RecurringGiftPayment') {
-    giftData.linked_gifts = payload.linked_gifts
-  }
-
-  // create gift
-  await blackbaudSkyApiClient.createGift(giftData)
-
-  return
+  return blackbaudSkyApiClient.createGift(giftData)
 }
 
 const action: ActionDefinition<Settings, Payload> = {
